Mechanism-Aware imputation.

A function to replace NA in the data frame by MAI::MAI() with random forest and single imputation.

Usage

impute_mai(missdf, ...)

Arguments

missdf

a data frame with missing values to be imputed containing features in columns and samples in rows.

...

other parameters of MAI::MAI() besides data_miss.

Value

A data.frame with imputed values by MAI::MAI().

Silent defaults

MCAR_algorithm is set to random_forest, MNAR_algorithm is set to single and verbose is set to FALSE.

References

Dekermanjian JP, Shaddox E, Nandy D, Ghosh D, Kechris K (2022). “Mechanism-Aware Imputation: A Two-Step Approach in Handling Missing Values in Metabolomics.” BMC Bioinformatics, 23(1), 179. ISSN 1471-2105, doi:10.1186/s12859-022-04659-1 .

See also

Examples

data(sim_miss_large)
impute_mai(sim_miss_large)
#>            X1         X2          X3          X4         X5         X6
#> 1  0.71591583 0.51676348 0.255616798 0.072616067 0.47282582 0.21518149
#> 2  0.07824958 0.85780800 0.513020133 0.522022313 0.11693067 0.71989643
#> 3  0.81128719 0.35854257 0.508994550 0.163078383 0.80347375 0.02330026
#> 4  0.08372887 0.65613454 0.114844928 0.676851177 0.45600733 0.89491071
#> 5  0.57447183 0.77252534 0.288813056 0.970359302 0.31999267 0.78093747
#> 6  0.09149598 0.42835757 0.701997595 0.003004555 0.44996093 0.36460359
#> 7  0.57935689 0.60706762 0.578279944 0.760383036 0.45767100 0.50590879
#> 8  0.62149079 0.83034611 0.660347324 0.414356061 0.46061268 0.47941773
#> 9  0.65428158 0.44310293 0.474469531 0.450800542 0.18936341 0.49319117
#> 10 0.58585609 0.44698476 0.371565553 0.833527265 0.19727859 0.36275295
#> 11 0.27936977 0.07313514 0.865470098 0.468887639 0.69936970 0.21051518
#> 12 0.54886031 0.46079630 0.490285160 0.436025625 0.02405465 0.03680037
#> 13 0.59763514 0.09559947 0.435709775 0.466615232 0.53110866 0.04339994
#> 14 0.49508477 0.88323422 0.524291929 0.463485106 0.55174990 0.54574019
#> 15 0.40332773 0.64798243 0.505505762 0.216255357 0.30242229 0.61461557
#> 16 0.45144701 0.47544209 0.513155607 0.422483300 0.52021335 0.63781590
#> 17 0.25428257 0.87782877 0.925002274 0.150749966 0.91212385 0.91801071
#> 18 0.44010786 0.01197316 0.618232834 0.463135916 0.84778290 0.49810895
#> 19 0.98671629 0.18128744 0.642066261 0.553355878 0.86023184 0.08542084
#> 20 0.56249247 0.83463910 0.863103885 0.321668908 0.54402641 0.79037477
#> 21 0.53145688 0.45516802 0.705921304 0.611937849 0.55219503 0.40208819
#> 22 0.91942905 0.55920404 0.815114947 0.273988729 0.45697717 0.55486997
#> 23 0.46832531 0.54809508 0.461234877 0.978713853 0.98445131 0.54296369
#> 24 0.89967340 0.85441060 0.432035017 0.582151069 0.91723185 0.34024228
#> 25 0.53719482 0.18905533 0.004868873 0.537738671 0.31444418 0.16824111
#> 26 0.58083342 0.26157128 0.331417043 0.030991118 0.83971687 0.99814141
#> 27 0.68378756 0.68071316 0.274083114 0.651652732 0.08575918 0.18961351
#> 28 0.38790363 0.14553118 0.588681038 0.420136203 0.42434504 0.55099109
#> 29 0.36750876 0.92638072 0.738694755 0.479462505 0.08303265 0.81817688
#> 30 0.21713883 0.92382190 0.511686071 0.856528853 0.33873010 0.21169651
#> 31 0.05472327 0.16163229 0.698022422 0.249073416 0.84668550 0.33416518
#> 32 0.35042474 0.03001550 0.888251950 0.608276315 0.46181259 0.83595666
#> 33 0.81518507 0.45370422 0.503436295 0.489609561 0.18974039 0.49556043
#> 34 0.58474697 0.67492179 0.215658443 0.131562222 0.81824357 0.46237572
#> 35 0.92087373 0.42850297 0.497280255 0.484943496 0.93749363 0.31827936
#> 36 0.31096283 0.46474583 0.106057892 0.723440063 0.54978647 0.52403136
#> 37 0.82539084 0.49664549 0.464715083 0.442366773 0.78061133 0.93128853
#> 38 0.81396203 0.73386703 0.924325841 0.490951507 0.47055501 0.15100504
#> 39 0.94358824 0.64778917 0.942146245 0.519850599 0.59156763 0.36522315
#> 40 0.97450871 0.56742042 0.109282812 0.947927148 0.84164652 0.35708874
#> 41 0.19797690 0.57366817 0.423260428 0.597773543 0.89840414 0.15624692
#> 42 0.43792501 0.41735240 0.452208367 0.521543029 0.21256088 0.25854356
#> 43 0.16207378 0.07485813 0.310625761 0.616323057 0.56789714 0.48146764
#> 44 0.63533143 0.44695072 0.984235951 0.875356896 0.25004764 0.11077702
#> 45 0.82875620 0.51192572 0.706515786 0.257768207 0.15346264 0.56036121
#> 46 0.99848865 0.78621309 0.312068742 0.349681286 0.58205616 0.94107206
#> 47 0.97945078 0.82159823 0.527049262 0.300909912 0.45334876 0.42121408
#> 48 0.86979109 0.14573593 0.321930339 0.668125439 0.51313580 0.50515961
#> 49 0.85244770 0.99754286 0.836515941 0.507602901 0.10946054 0.44886246
#> 50 0.55420554 0.47867220 0.075014113 0.582006406 0.56338573 0.47229553
#>            X7          X8         X9        X10         X11        X12
#> 1  0.25014984 0.346437299 0.34002227 0.81559545 0.235712131 0.88027411
#> 2  0.55018359 0.944148119 0.80070800 0.40460324 0.531016889 0.40288868
#> 3  0.51826520 0.870752740 0.32737357 0.74594254 0.849290876 0.96820083
#> 4  0.49389966 0.289229429 0.96630209 0.54455675 0.612319493 0.11642238
#> 5  0.40752016 0.630247698 0.10603724 0.35399003 0.346325504 0.44738517
#> 6  0.64751159 0.969346175 0.29213213 0.76878302 0.906982573 0.53933339
#> 7  0.26462930 0.827610224 0.69063195 0.53276770 0.169867788 0.04950492
#> 8  0.04819104 0.637177039 0.38651752 0.25020823 0.404862234 0.48395048
#> 9  0.49772583 0.080198513 0.83009657 0.14219668 0.140922151 0.92718222
#> 10 0.26026490 0.993168920 0.64874084 0.54196154 0.672286163 0.51375066
#> 11 0.48583931 0.641506840 0.53008613 0.57233867 0.928711836 0.53963774
#> 12 0.13261942 0.115893963 0.54425212 0.65250981 0.888750179 0.41963473
#> 13 0.47044397 0.561182958 0.48623829 0.12634779 0.513998714 0.11710036
#> 14 0.88737273 0.520395608 0.84020446 0.10065341 0.689324722 0.06873683
#> 15 0.04176691 0.414728324 0.85947040 0.24003759 0.754697058 0.10344070
#> 16 0.39007090 0.426839811 0.23840381 0.84476454 0.597073818 0.51705127
#> 17 0.35511818 0.540626562 0.61806642 0.54043940 0.942157283 0.09536647
#> 18 0.99439918 0.080021261 0.39672232 0.47211097 0.717251634 0.02953261
#> 19 0.48338353 0.413585206 0.65056256 0.73831190 0.541126573 0.44946897
#> 20 0.84857426 0.804817031 0.39922167 0.09051804 0.643970801 0.77005794
#> 21 0.20935890 0.588335072 0.86078490 0.56508407 0.582962100 0.35558199
#> 22 0.55265940 0.533543820 0.51003730 0.75636365 0.008220428 0.84781157
#> 23 0.21179588 0.417310277 0.39862106 0.44827008 0.574670374 0.48972306
#> 24 0.15933349 0.691476746 0.41447241 0.40067600 0.156315880 0.83229332
#> 25 0.64153720 0.498119849 0.38202505 0.58049706 0.443925294 0.64851879
#> 26 0.44986095 0.435089825 0.52365691 0.72500804 0.503634276 0.42413138
#> 27 0.97219152 0.526337931 0.16210898 0.34914926 0.202769139 0.54781051
#> 28 0.52731441 0.482998525 0.96426670 0.18850780 0.825465299 0.40592829
#> 29 0.57257807 0.386756063 0.86783554 0.72671069 0.791072183 0.51228306
#> 30 0.85346947 0.584205846 0.27398992 0.21403068 0.593612797 0.62757755
#> 31 0.63437011 0.534098086 0.50594732 0.92338850 0.767464769 0.61848689
#> 32 0.84473953 0.099341821 0.57638614 0.33945099 0.797799595 0.14542157
#> 33 0.47064513 0.091276783 0.75367169 0.37183663 0.003183017 0.63115770
#> 34 0.10988111 0.987992536 0.81453959 0.12315529 0.812321605 0.47755882
#> 35 0.23568693 0.473658998 0.82298732 0.12273551 0.252322710 0.16291966
#> 36 0.99790850 0.414806289 0.43644185 0.48127874 0.526955199 0.84107165
#> 37 0.57787007 0.215245785 0.78801446 0.11947607 0.221781736 0.30619064
#> 38 0.02067985 0.816517546 0.05574163 0.69315854 0.657629620 0.51329498
#> 39 0.42323795 0.416511635 0.15897616 0.91785010 0.996626715 0.84557507
#> 40 0.60669667 0.003039246 0.18694749 0.02764557 0.557215076 0.62268345
#> 41 0.56276462 0.593277852 0.65399859 0.69479450 0.568829441 0.46356683
#> 42 0.72522007 0.573210305 0.65086182 0.20477864 0.632455572 0.14853009
#> 43 0.43674910 0.440599329 0.85423062 0.43155598 0.577322546 0.17596321
#> 44 0.83018503 0.871899163 0.27901016 0.80773845 0.313355858 0.49498268
#> 45 0.55599695 0.476764283 0.09650713 0.38759558 0.005396917 0.42874437
#> 46 0.69114037 0.375669455 0.46703653 0.31751918 0.173270121 0.81322514
#> 47 0.26008379 0.799240797 0.42122160 0.45002499 0.126809106 0.09403678
#> 48 0.29453751 0.369274247 0.05789243 0.35523492 0.952763613 0.99768084
#> 49 0.41795816 0.954226248 0.27765898 0.04595447 0.429080302 0.53311968
#> 50 0.97258815 0.569616268 0.39988880 0.99265020 0.965724215 0.92305343
#>           X13        X14        X15        X16        X17        X18        X19
#> 1  0.65779540 0.99258196 0.23228185 0.49581666 0.77078893 0.58775489 0.13178215
#> 2  0.53047017 0.72006415 0.99056785 0.61382006 0.05461837 0.07083153 0.14078500
#> 3  0.76169227 0.84851607 0.74482818 0.16373791 0.51676101 0.43919116 0.37895652
#> 4  0.07285081 0.80800678 0.60549785 0.67134500 0.61790795 0.62912677 0.45979413
#> 5  0.61880946 0.52996889 0.10949635 0.34541930 0.54404343 0.98550523 0.51282003
#> 6  0.54308043 0.13835160 0.05609576 0.81927333 0.14801321 0.87761138 0.88966380
#> 7  0.21562248 0.65415144 0.22936834 0.51048782 0.98183092 0.32856767 0.71754346
#> 8  0.91996728 0.82404509 0.25692058 0.49342217 0.06902123 0.12870940 0.98035214
#> 9  0.35350501 0.64051000 0.18593751 0.46768835 0.66491502 0.07966607 0.49287341
#> 10 0.43401900 0.55011139 0.95669033 0.50602917 0.73519518 0.91750256 0.38282238
#> 11 0.44424450 0.91171850 0.25880243 0.35846521 0.58807076 0.47084184 0.52942790
#> 12 0.12130476 0.03107148 0.84870539 0.51123625 0.23727297 0.50485443 0.55593724
#> 13 0.47797038 0.94438747 0.90314025 0.55709743 0.95258779 0.45184035 0.65021986
#> 14 0.44976909 0.62369878 0.85171827 0.47550759 0.95799332 0.82399074 0.48356143
#> 15 0.24376562 0.94135472 0.08707751 0.18449129 0.45567284 0.83437854 0.53620792
#> 16 0.80108546 0.71224208 0.49979076 0.52281632 0.57111784 0.47194716 0.97149862
#> 17 0.47062246 0.17154510 0.46054528 0.70433645 0.28719868 0.53489234 0.26097238
#> 18 0.13535990 0.19368219 0.94092586 0.59695156 0.64926408 0.69197731 0.50322229
#> 19 0.36521766 0.01757208 0.44696275 0.46367131 0.23105947 0.51764344 0.37933235
#> 20 0.86698592 0.54044143 0.02323549 0.60009956 0.41696830 0.49864654 0.27041921
#> 21 0.56956569 0.50095671 0.59990758 0.36858896 0.99754067 0.43202602 0.54546437
#> 22 0.95554486 0.46005325 0.43982579 0.89793371 0.81738246 0.92179669 0.90783105
#> 23 0.73369680 0.39071023 0.95514669 0.51500156 0.41022313 0.80853598 0.39483029
#> 24 0.59013495 0.62100745 0.96095461 0.55042319 0.04572675 0.14755090 0.26436742
#> 25 0.47497122 0.09661039 0.11108629 0.90132959 0.15259932 0.40156217 0.18716397
#> 26 0.92407550 0.02091660 0.15692117 0.92602551 0.84003871 0.51431236 0.42987282
#> 27 0.38730003 0.64735652 0.50066462 0.10278319 0.97942372 0.74082050 0.75538807
#> 28 0.08599180 0.35265585 0.07267714 0.68114207 0.49933355 0.17993548 0.54606533
#> 29 0.50992305 0.52886027 0.78512356 0.59368917 0.59220568 0.51994820 0.44827658
#> 30 0.56926426 0.81247856 0.34579784 0.12982113 0.44646032 0.79429046 0.53298141
#> 31 0.33117311 0.92097885 0.96561575 0.33662531 0.66100445 0.52079944 0.51151257
#> 32 0.63514498 0.48538071 0.52384694 0.68611093 0.95856579 0.60143524 0.39916585
#> 33 0.74156237 0.49634916 0.58624759 0.92776881 0.60966435 0.50862623 0.74740614
#> 34 0.50508957 0.70988232 0.14556295 0.14939888 0.51466560 0.52085626 0.76164536
#> 35 0.41787749 0.49197271 0.13187634 0.48246416 0.46618891 0.20138771 0.44439002
#> 36 0.52169512 0.57529934 0.74830484 0.42951248 0.62549486 0.89498532 0.45215219
#> 37 0.51304275 0.48403251 0.83176413 0.18300401 0.72275398 0.73963882 0.51144165
#> 38 0.30936912 0.61166596 0.00422933 0.13056167 0.67048131 0.78621964 0.08652295
#> 39 0.51342601 0.02750766 0.43048408 0.51106758 0.35656082 0.49419361 0.09401477
#> 40 0.33744518 0.59442009 0.94201231 0.92220980 0.94010209 0.40591905 0.01752804
#> 41 0.90073927 0.97988650 0.67642149 0.55751435 0.96798727 0.55434164 0.73993635
#> 42 0.48281658 0.71737293 0.98264524 0.93503855 0.35741287 0.46744561 0.80979039
#> 43 0.29014090 0.50465975 0.29320738 0.51962811 0.54613758 0.45779591 0.16148632
#> 44 0.20444900 0.68046129 0.91568144 0.05784743 0.50097939 0.16303349 0.41514653
#> 45 0.55796272 0.30087972 0.21158127 0.63366622 0.13919866 0.35586901 0.44403699
#> 46 0.91000861 0.90813959 0.43108651 0.26827598 0.01323098 0.47810067 0.68004003
#> 47 0.51948087 0.79174892 0.56396509 0.44508025 0.79664452 0.87395240 0.80891265
#> 48 0.43938447 0.50682834 0.44192820 0.18720365 0.18965469 0.46841693 0.21566916
#> 49 0.53880730 0.71706740 0.49454358 0.52313993 0.07193703 0.20206301 0.82726126
#> 50 0.51887226 0.98520012 0.63237771 0.49972848 0.86768813 0.54534957 0.19349583
#>           X20        X21         X22        X23        X24        X25
#> 1  0.82876940 0.18480442 0.785232264 0.93434067 0.40905508 0.77922642
#> 2  0.04351867 0.58605523 0.670590825 0.09505949 0.27345575 0.34897139
#> 3  0.49260385 0.77242107 0.039911876 0.33064559 0.40270486 0.06427560
#> 4  0.54671715 0.32134943 0.774555078 0.95740765 0.64422276 0.96133653
#> 5  0.75263971 0.25061276 0.993973556 0.51051750 0.53171116 0.62301472
#> 6  0.88630574 0.87897764 0.971536968 0.57985483 0.13495040 0.38776695
#> 7  0.85462470 0.63687167 0.438176297 0.34572796 0.88771491 0.64063137
#> 8  0.84022547 0.50488876 0.560700127 0.40284294 0.32511595 0.22877823
#> 9  0.39534133 0.82752554 0.333145374 0.83731627 0.57057424 0.65234741
#> 10 0.56324610 0.98423004 0.965220488 0.32621758 0.23276455 0.03441201
#> 11 0.51067152 0.73919520 0.479713262 0.90898888 0.52333634 0.78078729
#> 12 0.90468534 0.52593333 0.050114230 0.16624970 0.32206083 0.39244659
#> 13 0.57793110 0.96515268 0.244144144 0.51092341 0.23665062 0.43320871
#> 14 0.20647782 0.32102240 0.144263827 0.14308706 0.38413775 0.26694607
#> 15 0.68350813 0.53091220 0.443981157 0.62770529 0.88579247 0.51144693
#> 16 0.41606841 0.29626291 0.488004518 0.61301753 0.58515118 0.60374918
#> 17 0.43543898 0.21741945 0.055802328 0.47775057 0.55675469 0.58430021
#> 18 0.28282603 0.51281106 0.426990480 0.53547069 0.03866548 0.32505318
#> 19 0.11759215 0.47983052 0.019758131 0.93196728 0.47531932 0.92168931
#> 20 0.32975324 0.81162349 0.499512185 0.16830807 0.17674116 0.57455985
#> 21 0.73843596 0.59194853 0.592455837 0.57383656 0.39129349 0.61700382
#> 22 0.70132192 0.31323200 0.634254814 0.63945309 0.75788357 0.88419571
#> 23 0.48318659 0.92521622 0.507153545 0.49473058 0.32789253 0.51850242
#> 24 0.56302559 0.82046467 0.958107898 0.58280263 0.48952433 0.53527950
#> 25 0.36086808 0.58058237 0.778765526 0.32597036 0.52330707 0.41273847
#> 26 0.16338780 0.78740063 0.022452049 0.47521219 0.61094976 0.57891047
#> 27 0.84892145 0.47265874 0.364679370 0.18631976 0.72388038 0.42760263
#> 28 0.31434533 0.69477169 0.688753319 0.62796334 0.45352252 0.87756794
#> 29 0.92045476 0.12440047 0.691509332 0.81008120 0.16910886 0.35023866
#> 30 0.98788191 0.66633700 0.921728698 0.52069517 0.51212623 0.49870671
#> 31 0.41137154 0.28024400 0.463907293 0.73714904 0.47094081 0.62080363
#> 32 0.45754834 0.54987309 0.426591306 0.49318281 0.65418499 0.26864363
#> 33 0.53903690 0.93821390 0.152655850 0.77891620 0.59330054 0.43794484
#> 34 0.55545249 0.17235720 0.146975832 0.45324014 0.46067249 0.59240097
#> 35 0.14280554 0.58383031 0.410289504 0.10111136 0.34521707 0.55188351
#> 36 0.04756337 0.99218365 0.494525188 0.09974921 0.72890748 0.04243355
#> 37 0.41376424 0.56773387 0.111733630 0.37396452 0.94684467 0.56068784
#> 38 0.13054721 0.50115041 0.482498099 0.73936156 0.67142658 0.41168071
#> 39 0.43310340 0.03693727 0.411757322 0.64935269 0.86538002 0.86497099
#> 40 0.07067445 0.04587781 0.358508810 0.87471499 0.55045768 0.95618982
#> 41 0.69914571 0.58665418 0.633321723 0.10615417 0.96215071 0.52622938
#> 42 0.78374680 0.60994111 0.905436506 0.90741657 0.57161547 0.36834152
#> 43 0.40113046 0.44218804 0.621198986 0.57792724 0.14444661 0.99345568
#> 44 0.61055951 0.27300412 0.004944553 0.89027505 0.53564952 0.74465078
#> 45 0.21544088 0.49698089 0.398942701 0.27522827 0.92906105 0.26786195
#> 46 0.51323710 0.30545882 0.339215387 0.81381360 0.14954124 0.62944478
#> 47 0.84055003 0.03110503 0.183297911 0.08588486 0.31085753 0.03830732
#> 48 0.29492813 0.58514395 0.476246926 0.42094921 0.27948596 0.51162911
#> 49 0.77636455 0.58709857 0.317538884 0.35790986 0.55785816 0.94530287
#> 50 0.41854690 0.44820061 0.247421121 0.91698722 0.47679345 0.71641699
#>           X26        X27        X28        X29        X30        X31        X32
#> 1  0.30797190 0.46447436 0.16893035 0.62317594 0.37351537 0.42104403 0.60040097
#> 2  0.22131468 0.66506163 0.56737316 0.01733979 0.49390193 0.51488889 0.14811102
#> 3  0.64031056 0.19934907 0.70926339 0.49378767 0.59302483 0.44560135 0.52513402
#> 4  0.48158825 0.35075799 0.84699845 0.70118599 0.54537081 0.50696413 0.64453902
#> 5  0.51718171 0.75915348 0.47450290 0.49002940 0.23429647 0.24544435 0.34890557
#> 6  0.96353752 0.58217807 0.12434160 0.50181105 0.76183626 0.14402755 0.44843037
#> 7  0.28298504 0.87401360 0.74492913 0.20875312 0.02464244 0.02726733 0.44397119
#> 8  0.83410436 0.58739520 0.48866414 0.26563438 0.93769541 0.42682825 0.30737200
#> 9  0.91197361 0.14675782 0.73023000 0.34577651 0.87345798 0.96537949 0.20153060
#> 10 0.97041437 0.51138209 0.02255707 0.15953522 0.22936393 0.24213256 0.48312912
#> 11 0.21287050 0.25056027 0.52091640 0.97200577 0.50538652 0.58809276 0.76990571
#> 12 0.56723527 0.81450309 0.34573461 0.80951917 0.93579854 0.65151129 0.87121402
#> 13 0.06248995 0.66088589 0.06453708 0.62237228 0.97883613 0.06022198 0.94464231
#> 14 0.66309327 0.29591748 0.38490453 0.58840923 0.64320687 0.49353920 0.81113681
#> 15 0.64267090 0.74647972 0.83197151 0.22050847 0.46537814 0.41277221 0.32951492
#> 16 0.29000896 0.98396537 0.44288637 0.78234104 0.46246696 0.97486470 0.53988129
#> 17 0.66527857 0.43684671 0.46740604 0.95337167 0.29557299 0.50057992 0.22468124
#> 18 0.87087930 0.41478457 0.93706709 0.65754500 0.95010951 0.65202679 0.71614224
#> 19 0.73799448 0.45625595 0.85047318 0.64095359 0.49414662 0.85753843 0.40006517
#> 20 0.61834867 0.18856722 0.80812561 0.00644259 0.84524527 0.53883613 0.45948536
#> 21 0.29804202 0.32175938 0.02169952 0.45189417 0.68010683 0.06102330 0.70704788
#> 22 0.91175667 0.83417756 0.14459028 0.75291747 0.94248849 0.27868585 0.02824067
#> 23 0.74689911 0.13427655 0.75699591 0.34362774 0.29434395 0.88272130 0.80139249
#> 24 0.90791342 0.25051908 0.46564279 0.29649668 0.24411543 0.42834299 0.07901599
#> 25 0.68093261 0.31619363 0.61905019 0.90490183 0.97441046 0.30358258 0.64165734
#> 26 0.47589680 0.65122221 0.56817065 0.00850901 0.43196017 0.57949865 0.74432785
#> 27 0.82501754 0.22826261 0.92735329 0.94228691 0.57784712 0.33331427 0.26542311
#> 28 0.58548276 0.12845060 0.04988626 0.71396794 0.59638814 0.55159087 0.47613053
#> 29 0.31383689 0.96926197 0.76179000 0.93765134 0.50203177 0.17064776 0.16070765
#> 30 0.76825267 0.02920483 0.51925526 0.40791722 0.90591492 0.45663301 0.46069966
#> 31 0.53585358 0.28839672 0.87364853 0.83522704 0.01418827 0.54672440 0.27520758
#> 32 0.66810154 0.29591652 0.46787733 0.86662406 0.47758592 0.39505911 0.51603923
#> 33 0.95298304 0.68338133 0.33962855 0.44131155 0.02162108 0.01459659 0.34072923
#> 34 0.54573305 0.43856386 0.34964633 0.47645191 0.93720302 0.08582418 0.32602015
#> 35 0.92240082 0.25508552 0.77629677 0.65563874 0.09629750 0.42944033 0.45485591
#> 36 0.93792138 0.72649314 0.28590024 0.46153491 0.59112517 0.33754472 0.35185131
#> 37 0.84698802 0.77787109 0.32605117 0.46568318 0.21315395 0.98723910 0.23522838
#> 38 0.51144044 0.51351024 0.48825537 0.05818326 0.29984047 0.02505622 0.91335526
#> 39 0.98175084 0.37102296 0.27265322 0.89504648 0.46770093 0.93179930 0.82046400
#> 40 0.87897639 0.17178284 0.92007517 0.54479457 0.13444687 0.81148375 0.30680695
#> 41 0.52080782 0.19613550 0.77614683 0.02963982 0.44293095 0.85208246 0.82687756
#> 42 0.42618924 0.99619631 0.52346166 0.80169799 0.59032273 0.86005031 0.55419816
#> 43 0.54187741 0.35028094 0.15013351 0.89755971 0.47774847 0.46220331 0.50403493
#> 44 0.18345732 0.58672025 0.52761634 0.52658572 0.22485303 0.28564828 0.20142814
#> 45 0.26679883 0.66541183 0.67795750 0.36245032 0.25119635 0.41633178 0.34443818
#> 46 0.67227211 0.16708576 0.75872796 0.69917647 0.63591440 0.60083342 0.50643583
#> 47 0.24142119 0.60659383 0.19827383 0.29098102 0.50979300 0.40563077 0.17976516
#> 48 0.59328050 0.50446173 0.20585577 0.75829686 0.51967562 0.23021446 0.29454116
#> 49 0.77070232 0.45456981 0.07118774 0.44105097 0.06570891 0.72150475 0.03341543
#> 50 0.35303164 0.75064232 0.98596582 0.51867606 0.56861898 0.93767767 0.59020977
#>            X33         X34          X35        X36        X37         X38
#> 1  0.511683520 0.345173705 0.6247019173 0.46190493 0.43253326 0.115577885
#> 2  0.584865713 0.220702644 0.6273962804 0.26462937 0.94749960 0.789506597
#> 3  0.037524378 0.486014152 0.4133258513 0.86867765 0.60806996 0.797133448
#> 4  0.569726389 0.588867821 0.9207097208 0.36489641 0.80136494 0.550081308
#> 5  0.005454513 0.384689845 0.8759513139 0.69186239 0.12028763 0.286285537
#> 6  0.264961040 0.498186709 0.5504697615 0.99000248 0.41461827 0.831583387
#> 7  0.631994247 0.602239255 0.9741398243 0.21038604 0.94630702 0.842768101
#> 8  0.654308783 0.574458848 0.3237321766 0.35576712 0.54994287 0.545919008
#> 9  0.786054220 0.493637608 0.2737691046 0.57609463 0.53941200 0.901376349
#> 10 0.261410854 0.391217505 0.7134763140 0.94933662 0.94353614 0.173843937
#> 11 0.053384619 0.180501444 0.4051274641 0.74547898 0.72199926 0.056073593
#> 12 0.907679759 0.604084631 0.7322553063 0.39812264 0.54573202 0.086119739
#> 13 0.049013627 0.787866242 0.6243370529 0.74689181 0.84615525 0.445014183
#> 14 0.633977301 0.345192742 0.7390514382 0.02463594 0.10195317 0.645468470
#> 15 0.878673413 0.168064992 0.1634237682 0.06033942 0.51468680 0.612942996
#> 16 0.453061693 0.860310045 0.8477109061 0.45706448 0.67838289 0.639810285
#> 17 0.227331186 0.464648673 0.5268927582 0.10078785 0.92750830 0.167456558
#> 18 0.163518806 0.696137673 0.7136495339 0.55131341 0.28548507 0.671444730
#> 19 0.987583713 0.557954967 0.7382442530 0.50189803 0.62572096 0.555257898
#> 20 0.516084441 0.332727081 0.1663744955 0.14227464 0.58100900 0.519926402
#> 21 0.265436746 0.465733482 0.6091289739 0.87738118 0.98107209 0.875225405
#> 22 0.144266190 0.982968777 0.5612601175 0.49667675 0.01779024 0.868978215
#> 23 0.465053277 0.434288519 0.9942641135 0.20010241 0.60560775 0.934411318
#> 24 0.321048825 0.470128338 0.0267528780 0.58070688 0.54092015 0.154518418
#> 25 0.316327626 0.449342496 0.6058761158 0.85261944 0.59893722 0.246503730
#> 26 0.909100895 0.684089508 0.8838235207 0.33153402 0.03624712 0.642551303
#> 27 0.525820237 0.725241523 0.1531594398 0.26101761 0.68857410 0.921900905
#> 28 0.971836349 0.324584927 0.2078533291 0.18805179 0.88345675 0.728297831
#> 29 0.530064316 0.851595982 0.8507306906 0.12229006 0.55095503 0.985584209
#> 30 0.689577265 0.531590016 0.4838525266 0.11156233 0.78577483 0.870621627
#> 31 0.118125303 0.506448993 0.8191220823 0.32908460 0.93242776 0.329818901
#> 32 0.347948819 0.813023071 0.1636604709 0.62090924 0.89043720 0.749487144
#> 33 0.047300944 0.580895690 0.3161884279 0.50806906 0.93496005 0.724960342
#> 34 0.354738628 0.017664609 0.2689461680 0.90109805 0.31199840 0.364452341
#> 35 0.124288302 0.456692807 0.7143825532 0.97424371 0.42144254 0.526135030
#> 36 0.732160954 0.137529943 0.3443183557 0.42237275 0.45847286 0.878043310
#> 37 0.321412294 0.433703340 0.4888231969 0.03531950 0.41933633 0.573053232
#> 38 0.170222537 0.084301427 0.5662569541 0.11707190 0.63643410 0.009890485
#> 39 0.612084733 0.929803638 0.9860978583 0.39041585 0.85331411 0.701161839
#> 40 0.442296885 0.544805619 0.4492896605 0.86180699 0.82113030 0.538588945
#> 41 0.480091607 0.400232798 0.2029350933 0.06893562 0.95034001 0.567383993
#> 42 0.269165098 0.004914962 0.5433851436 0.48961547 0.67658549 0.095071064
#> 43 0.847832166 0.496862229 0.5916413840 0.28956915 0.51907115 0.408151144
#> 44 0.727702224 0.449888946 0.8485561204 0.43627400 0.65755384 0.470009079
#> 45 0.915279565 0.563794985 0.7417069317 0.44347606 0.51661311 0.637979321
#> 46 0.028599375 0.115698863 0.0003303257 0.93254919 0.55337695 0.564848838
#> 47 0.047028878 0.808371289 0.5674606552 0.99359394 0.51360334 0.444271072
#> 48 0.891872453 0.296245911 0.9002497348 0.34991991 0.47325463 0.979995800
#> 49 0.032493589 0.611238711 0.4826963681 0.50811764 0.92440970 0.116307924
#> 50 0.418181912 0.770243284 0.6303533665 0.14282747 0.63516094 0.825470688
#>            X39        X40         X41        X42        X43        X44
#> 1  0.775906606 0.13737554 0.136672128 0.05878118 0.41822999 0.95972644
#> 2  0.385750609 0.62644279 0.476976337 0.45076891 0.18194533 0.51715852
#> 3  0.601551775 0.43694682 0.945313138 0.55489669 0.88397615 0.82352843
#> 4  0.841919949 0.53426175 0.026912781 0.48685560 0.09228840 0.42148682
#> 5  0.540120464 0.52395350 0.697623735 0.66953748 0.71957070 0.74342039
#> 6  0.391530135 0.55608117 0.473139671 0.97221802 0.89438287 0.88740367
#> 7  0.483377054 0.57516006 0.628119649 0.58699202 0.52161819 0.58006969
#> 8  0.482620504 0.59040768 0.223478626 0.58951136 0.98396314 0.50782467
#> 9  0.490498585 0.03622434 0.599874558 0.29593665 0.61520465 0.73284047
#> 10 0.696487672 0.83386371 0.477149399 0.98321255 0.02084732 0.90726488
#> 11 0.467202441 0.02876003 0.532988219 0.54613662 0.57794125 0.29818485
#> 12 0.004370333 0.60902048 0.644755020 0.83154281 0.46286208 0.60337333
#> 13 0.208885702 0.85706098 0.954726063 0.25343885 0.30291842 0.09773602
#> 14 0.738501565 0.58784368 0.690975328 0.98861007 0.44044074 0.51936911
#> 15 0.107413489 0.86104101 0.474541905 0.54573542 0.17946257 0.39317358
#> 16 0.496171475 0.57006581 0.207459769 0.11650241 0.73490443 0.37036374
#> 17 0.969621916 0.64095132 0.571558677 0.60632678 0.64497067 0.53508370
#> 18 0.843962974 0.78849683 0.207098268 0.50113993 0.24340254 0.49185295
#> 19 0.701543136 0.91787840 0.474354975 0.60786286 0.46318583 0.75517329
#> 20 0.591941979 0.28180881 0.795014372 0.54836290 0.27433853 0.93942242
#> 21 0.940373796 0.49452236 0.671667575 0.58212786 0.52223020 0.84682998
#> 22 0.389963259 0.86302188 0.498389067 0.56452106 0.63443377 0.36987487
#> 23 0.896749785 0.85278747 0.538889669 0.54849936 0.47185642 0.53794187
#> 24 0.943253464 0.43640576 0.810302546 0.84565150 0.87169486 0.58273719
#> 25 0.383570795 0.03090675 0.537472821 0.81872441 0.68168362 0.13435388
#> 26 0.415980124 0.56857833 0.964448194 0.60082437 0.58692350 0.44271235
#> 27 0.811047234 0.29574625 0.599128756 0.31782630 0.88849558 0.49118773
#> 28 0.635152133 0.02864544 0.548827875 0.21582287 0.24841314 0.20361642
#> 29 0.475169923 0.71828360 0.483874036 0.52760374 0.78320412 0.48800782
#> 30 0.258263598 0.32855097 0.577634003 0.27587187 0.70241251 0.53557753
#> 31 0.447523895 0.52088897 0.120589416 0.64032424 0.90543728 0.60019435
#> 32 0.785832142 0.23841525 0.753571586 0.47066212 0.78803241 0.47227734
#> 33 0.508019527 0.30237665 0.601394360 0.93735590 0.10490696 0.30628844
#> 34 0.142563991 0.07329489 0.881001580 0.68041778 0.06368659 0.88686236
#> 35 0.422801445 0.94091232 0.001599045 0.57539213 0.55719022 0.56689036
#> 36 0.923462192 0.94134256 0.481728495 0.13655585 0.42983120 0.36790599
#> 37 0.228234681 0.15096911 0.589570664 0.78478757 0.29296087 0.87401654
#> 38 0.518441714 0.86940943 0.478512920 0.51401831 0.87416357 0.22528805
#> 39 0.571019031 0.21515930 0.723592437 0.23648799 0.48073695 0.35898487
#> 40 0.589518817 0.38859365 0.680450574 0.99143036 0.52211829 0.31054141
#> 41 0.460389202 0.75525210 0.248712138 0.95143529 0.76390997 0.04299619
#> 42 0.074222219 0.83343274 0.514755329 0.34791016 0.54271836 0.53807337
#> 43 0.619938699 0.27957574 0.475824196 0.52643455 0.30235637 0.49849184
#> 44 0.326782173 0.75268899 0.312424511 0.56121887 0.65644203 0.51961088
#> 45 0.186631935 0.54498597 0.694063614 0.59328812 0.38978038 0.87964590
#> 46 0.522282130 0.32586336 0.605109787 0.43338342 0.49080145 0.55847399
#> 47 0.894184164 0.63651936 0.035894475 0.23437718 0.94416679 0.05148433
#> 48 0.884810792 0.90116262 0.371800276 0.49238632 0.16654968 0.53357891
#> 49 0.786750883 0.49408745 0.947606357 0.85038678 0.98724038 0.46493215
#> 50 0.094116013 0.73152306 0.560940019 0.47329467 0.23737475 0.77543568
#>           X45        X46         X47          X48        X49        X50
#> 1  0.85265667 0.59644738 0.852371223 0.0559872845 0.95735217 0.40804717
#> 2  0.21312367 0.71656146 0.848932273 0.8860864276 0.78142499 0.30112464
#> 3  0.85969451 0.49444693 0.352575504 0.2193175177 0.44458968 0.84646914
#> 4  0.83441649 0.80925188 0.773787788 0.3232960624 0.37811666 0.45806076
#> 5  0.66490328 0.53918535 0.202764235 0.1348798678 0.76385107 0.57020795
#> 6  0.12597194 0.61589529 0.365045791 0.0390430912 0.90004577 0.07107761
#> 7  0.85890424 0.82315470 0.445526347 0.0856559405 0.47699862 0.76780986
#> 8  0.65224769 0.86885817 0.330604936 0.5581821576 0.43177751 0.79552658
#> 9  0.55078225 0.46497566 0.394764594 0.5986114484 0.51362797 0.95230459
#> 10 0.47272634 0.15233776 0.482000188 0.7475190158 0.07477793 0.48588386
#> 11 0.82378212 0.58397434 0.371155697 0.5415068972 0.38673123 0.59567852
#> 12 0.64819584 0.37182793 0.093966853 0.4195758368 0.58047453 0.84906838
#> 13 0.67309519 0.53487434 0.300752096 0.0670880249 0.51839876 0.94932655
#> 14 0.13623939 0.46954354 0.565497953 0.1434748494 0.69330200 0.60685288
#> 15 0.25024253 0.62602582 0.806611250 0.9006192756 0.54391862 0.56751260
#> 16 0.72598149 0.24745054 0.435942930 0.2178305667 0.36846999 0.98707464
#> 17 0.11204385 0.37634032 0.438129374 0.9622041716 0.92661182 0.53312070
#> 18 0.08686740 0.79771808 0.354377773 0.4186276910 0.01686683 0.55070316
#> 19 0.32220006 0.67726721 0.018146015 0.5270466833 0.33777956 0.39078017
#> 20 0.49912608 0.92632645 0.094239566 0.4059477313 0.87798164 0.87669838
#> 21 0.99415064 0.19939128 0.265184810 0.3131020679 0.03389476 0.58968949
#> 22 0.28263819 0.99406671 0.478690606 0.4739536701 0.04974113 0.53231958
#> 23 0.47783679 0.54005964 0.408167861 0.6872922191 0.76789177 0.24686600
#> 24 0.89518913 0.84920749 0.803606101 0.2194121075 0.33643906 0.59192306
#> 25 0.41015335 0.58864746 0.361875111 0.4888826986 0.09173981 0.70801804
#> 26 0.51275618 0.66086618 0.612858488 0.6031443907 0.49027653 0.23031174
#> 27 0.49615836 0.74889546 0.994669823 0.9174510394 0.72037485 0.81680665
#> 28 0.57473867 0.93720755 0.457424424 0.1766561687 0.56111500 0.48593407
#> 29 0.08959223 0.84674301 0.841300351 0.4888186809 0.30422094 0.34069059
#> 30 0.46757242 0.61368799 0.502653239 0.8851070232 0.79786742 0.65159085
#> 31 0.70920223 0.91684349 0.392292177 0.1430236066 0.48082848 0.70688633
#> 32 0.08744409 0.65040369 0.036819978 0.1695205751 0.38676398 0.48327555
#> 33 0.75346902 0.82703753 0.348803202 0.4527747609 0.67645557 0.49371365
#> 34 0.64799628 0.34119660 0.300350453 0.6571703178 0.56467292 0.65511661
#> 35 0.48452660 0.35497377 0.445236275 0.7566334868 0.43214586 0.93047157
#> 36 0.45215102 0.41559419 0.246835504 0.5158763423 0.54845489 0.91755568
#> 37 0.49035837 0.54145828 0.076102891 0.2280937920 0.66545612 0.62557286
#> 38 0.61168985 0.24259184 0.158974137 0.3175933182 0.46027379 0.52719723
#> 39 0.54239400 0.91634121 0.491150573 0.5598035231 0.45206322 0.62321336
#> 40 0.31294629 0.57261208 0.540958368 0.9513210554 0.45419635 0.61469798
#> 41 0.52036397 0.52536575 0.003048244 0.6385835872 0.13170765 0.41611383
#> 42 0.49575855 0.57723157 0.154793021 0.2714969988 0.89350955 0.62463507
#> 43 0.63779757 0.85735466 0.463061271 0.9568858703 0.50315478 0.19844218
#> 44 0.75894519 0.07898811 0.464480039 0.5407590682 0.25187488 0.49943208
#> 45 0.47178580 0.90850509 0.528422663 0.3762958839 0.52991620 0.73620427
#> 46 0.34621478 0.88985470 0.491285231 0.9594069924 0.48732857 0.98342610
#> 47 0.94696829 0.69647444 0.092059590 0.0009776258 0.27524605 0.14008090
#> 48 0.62731761 0.25144984 0.393264137 0.4923162869 0.60018756 0.18976730
#> 49 0.53409648 0.55281155 0.571708690 0.5099724887 0.51386802 0.56576755
#> 50 0.52379049 0.91423906 0.673617329 0.5619869467 0.52810255 0.14123579
#>           X51        X52         X53        X54        X55         X56
#> 1  0.75392543 0.53624374 0.100945421 0.03827218 0.33866157 0.327068431
#> 2  0.12201413 0.51151379 0.008668519 0.71985147 0.88377183 0.899567314
#> 3  0.45585354 0.42894721 0.858034885 0.55977343 0.58373931 0.351186014
#> 4  0.47053737 0.41678973 0.801680101 0.63742933 0.42985437 0.563919633
#> 5  0.74452150 0.41901823 0.446510992 0.84395241 0.46488054 0.622621062
#> 6  0.95342100 0.18013605 0.005723736 0.31879973 0.53247777 0.472832702
#> 7  0.20977431 0.99788178 0.041260271 0.66456942 0.64087674 0.361458204
#> 8  0.60239821 0.51411588 0.073582107 0.69132171 0.87906586 0.573314751
#> 9  0.49927211 0.83690527 0.447845852 0.05559989 0.38229803 0.271067217
#> 10 0.35233271 0.46084937 0.448685671 0.33575825 0.23469915 0.610511969
#> 11 0.17383465 0.48749117 0.051918525 0.37605002 0.58335097 0.413970820
#> 12 0.84945249 0.55027167 0.516626204 0.74827247 0.44991725 0.566727434
#> 13 0.19002502 0.02992266 0.997684571 0.58304003 0.41543742 0.454815200
#> 14 0.87621609 0.69481353 0.289959890 0.63913834 0.68144195 0.303192276
#> 15 0.64603534 0.96114460 0.239719066 0.81931746 0.66262501 0.938709098
#> 16 0.51433306 0.69826076 0.919626287 0.84681290 0.41636213 0.430843226
#> 17 0.76879757 0.52408313 0.465911765 0.74146836 0.59648346 0.203775509
#> 18 0.80228014 0.94458187 0.333048016 0.59812652 0.56385937 0.306425831
#> 19 0.33330643 0.57043803 0.561337326 0.54131258 0.44657300 0.665730544
#> 20 0.66366510 0.38112801 0.855557301 0.92968727 0.32701864 0.775991111
#> 21 0.24079429 0.21890484 0.453164627 0.96555987 0.54649703 0.285927237
#> 22 0.93711769 0.75363977 0.240434051 0.49172896 0.53375826 0.605068816
#> 23 0.70331465 0.57835670 0.636735722 0.52724968 0.58224159 0.462465265
#> 24 0.58556264 0.25833174 0.798995158 0.74119437 0.46196346 0.722697668
#> 25 0.36882339 0.76390318 0.495313323 0.72214651 0.45639784 0.756776894
#> 26 0.63136296 0.36670907 0.059788672 0.20874384 0.90234275 0.463167882
#> 27 0.18656902 0.93842536 0.697347127 0.56486734 0.31477823 0.859960143
#> 28 0.60494013 0.34010487 0.933161369 0.80462821 0.40002247 0.320639248
#> 29 0.81707548 0.38792694 0.764271981 0.48862601 0.32802115 0.524983891
#> 30 0.36880450 0.53849091 0.468897349 0.23085770 0.74396844 0.955547050
#> 31 0.23760688 0.03871646 0.494428243 0.54695883 0.75636280 0.429093885
#> 32 0.51740203 0.54087561 0.426330769 0.62187553 0.61770666 0.009685838
#> 33 0.59630481 0.87975577 0.026264887 0.57710699 0.84846834 0.096605891
#> 34 0.52807770 0.47764874 0.782641638 0.75268438 0.49527277 0.468639730
#> 35 0.47022234 0.66257781 0.518914205 0.84359261 0.15893521 0.255464124
#> 36 0.48154209 0.65515263 0.953816800 0.63823565 0.87498921 0.415665249
#> 37 0.81799656 0.20604737 0.037709034 0.69748079 0.78848778 0.087805940
#> 38 0.02542528 0.20390679 0.096014467 0.24939612 0.84820158 0.406400577
#> 39 0.40140969 0.92778920 0.717028506 0.32318966 0.81769533 0.347442490
#> 40 0.62192897 0.91902570 0.378332425 0.94597710 0.34402963 0.374445478
#> 41 0.69702342 0.52048206 0.569281165 0.61472210 0.94798879 0.139168255
#> 42 0.24686927 0.09805039 0.500971340 0.58619891 0.08767204 0.843471515
#> 43 0.14082488 0.56989875 0.578266912 0.89665530 0.51052840 0.320111408
#> 44 0.41004618 0.25110303 0.388975086 0.54435601 0.72404297 0.714735691
#> 45 0.98144733 0.47401240 0.256660988 0.98788400 0.28911090 0.762047498
#> 46 0.31364259 0.62913537 0.747592290 0.57140748 0.09596265 0.363772867
#> 47 0.16421538 0.25547615 0.131087360 0.18106801 0.55128341 0.468986504
#> 48 0.47729443 0.17981645 0.685390384 0.40696933 0.68430815 0.522717367
#> 49 0.95773734 0.49958270 0.591920193 0.02950654 0.53584431 0.771844610
#> 50 0.52823634 0.81214103 0.789946306 0.23553105 0.50063615 0.396579811
#>            X57        X58        X59         X60
#> 1  0.607661297 0.55605174 0.67133232 0.847647345
#> 2  0.348056976 0.82883202 0.05434469 0.918572216
#> 3  0.348402343 0.17642022 0.17761943 0.041814897
#> 4  0.120870273 0.24134027 0.55256494 0.961467589
#> 5  0.696181237 0.76978535 0.77963056 0.166920776
#> 6  0.429253840 0.50838592 0.57174087 0.430531030
#> 7  0.419587293 0.97748136 0.82935558 0.479844628
#> 8  0.561913105 0.68326740 0.49093386 0.011140197
#> 9  0.727704766 0.43310330 0.85665688 0.675156984
#> 10 0.288967270 0.33103950 0.93706745 0.166264740
#> 11 0.692416366 0.68383684 0.20979381 0.029839897
#> 12 0.373611411 0.39609747 0.52966888 0.933460108
#> 13 0.578557200 0.34616837 0.91217421 0.468356296
#> 14 0.425914280 0.60071226 0.09532550 0.327545668
#> 15 0.705163931 0.36672712 0.50482359 0.196458016
#> 16 0.519202219 0.36155750 0.70128474 0.433665600
#> 17 0.596361426 0.31475508 0.55649674 0.538592321
#> 18 0.469350590 0.17018739 0.23861672 0.483758970
#> 19 0.050747627 0.33153594 0.55998359 0.555703374
#> 20 0.356909962 0.48927861 0.50501789 0.730212339
#> 21 0.082215664 0.99982463 0.09887524 0.366709199
#> 22 0.044953315 0.99199129 0.72117481 0.062374195
#> 23 0.667993196 0.21078561 0.89930506 0.925719552
#> 24 0.946572929 0.61434798 0.57196537 0.009948529
#> 25 0.928785255 0.50736106 0.37193030 0.505340631
#> 26 0.912150417 0.77973095 0.64821360 0.662025634
#> 27 0.120431869 0.85514083 0.39730657 0.740570713
#> 28 0.489481772 0.39992939 0.57768990 0.110968189
#> 29 0.815762093 0.53622156 0.47459618 0.280724511
#> 30 0.744249916 0.38149007 0.57733904 0.907026905
#> 31 0.796059325 0.10714858 0.93255369 0.661070938
#> 32 0.462246988 0.30612126 0.86717764 0.905815218
#> 33 0.203823405 0.07469455 0.22267029 0.114026087
#> 34 0.507730678 0.46196542 0.36958941 0.344720763
#> 35 0.706323777 0.54726161 0.65448698 0.424364945
#> 36 0.324051151 0.46236281 0.23047729 0.722886477
#> 37 0.004374429 0.27724019 0.49527577 0.170236649
#> 38 0.498014775 0.98481225 0.27615359 0.949438457
#> 39 0.992147781 0.71146446 0.67178543 0.646068114
#> 40 0.952730360 0.23716428 0.57834018 0.030827414
#> 41 0.707023354 0.90969052 0.22886175 0.783917933
#> 42 0.105780446 0.47632806 0.53943828 0.582030275
#> 43 0.700540401 0.84997866 0.67647149 0.764717694
#> 44 0.591389430 0.72475304 0.54552461 0.618889758
#> 45 0.008273253 0.30784014 0.56486168 0.788453106
#> 46 0.759317253 0.09596398 0.43322069 0.248626345
#> 47 0.443006099 0.78778857 0.27693308 0.705981409
#> 48 0.351030038 0.47868399 0.34222133 0.325582722
#> 49 0.389953071 0.62375038 0.89290295 0.130682606
#> 50 0.391128017 0.46626081 0.67624181 0.131953064