MetabImpute QRILC imputation.

A function to replace NA in the data frame by QRILC.

impute_metabimpute_qrilc(missdf, verbose = FALSE, ...)

Arguments

missdf

a data frame with missing values to be imputed containing features in columns and samples in rows.

verbose

boolean, if TRUE, prints the typical prompts of MetabImpute::Impute().

...

other parameters of MetabImpute::Impute() besides method and data.

Value

A data.frame with imputed values of QRILC by MetabImpute::Impute().

See also

Examples

data(sim_miss)
impute_metabimpute_qrilc(sim_miss)
#>             X1         X2         X3           X4          X5         X6
#> 1  0.155050833 0.88115677 0.10120640 0.1194841210 0.600405471 0.00000000
#> 2  0.968378809 0.60748407 0.63377903 0.6181194666 0.008132938 0.75207300
#> 3  0.468263086 0.57452956 0.24760893 0.0000000000 0.370125689 0.09063335
#> 4  0.776819652 0.80313329 0.55133933 0.1459775318 0.724248714 0.84777606
#> 5  0.407885741 0.79911692 0.23476108 0.1897135850 0.418270750 0.70873497
#> 6  0.538797149 0.80186265 0.25861469 0.5754945197 0.238248518 0.05553328
#> 7  0.830082966 0.75207300 0.95288810 0.0001881735 0.887881226 0.48833525
#> 8  0.187103555 0.57546133 0.85687457 0.3283282877 0.577775384 0.34722847
#> 9  0.779969688 0.94421829 0.02089214 0.0782780354 0.739950257 0.20304801
#> 10 0.193943927 0.21898736 0.55452418 0.6041721753 0.160679622 0.73995026
#> 11 0.434231178 0.47799791 0.87694381 0.3267990556 0.529571799 0.89369752
#> 12 0.002274518 0.04466879 0.80186265 0.7904525734 0.574529562 0.57452956
#> 13 0.834692139 0.16848571 0.54673734 0.8300829662 0.960086967 0.08444870
#> 14 0.000000000 0.36828087 0.23275320 0.8100510929 0.398616886 0.14825833
#> 15 0.956967818 0.56331137 0.60226778 0.2027530000 0.408468068 0.37028128
#> 16 0.948497073 0.20708728 0.71253911 0.5060450307 0.613321482 0.86117762
#> 17 0.600729976 0.73432757 0.64890360 0.5206679751 0.706590850 0.83668180
#> 18 0.261807405 0.70365677 0.18756242 0.0842680801 0.799116918 0.74054705
#> 19 0.643034251 0.73682406 0.23839289 0.2687653562 0.161331222 0.98576580
#> 20 0.526233507 0.06512834 0.00000000 0.8147157354 0.885932416 0.03208182