SoftImpute imputation.

A function to replace NA in the data frame by softImpute::softImpute().

Usage

impute_softimpute(missdf, ...)

Arguments

missdf

a data frame with missing values to be imputed containing features in columns and samples in rows.

...

other parameters of softImpute::softImpute() besides x.

Value

A data.frame with imputed values by softImpute::softImpute().

References

Mazumder TH, Rahul (2021). “softImpute: Matrix Completion via Iterative Soft-Thresholded SVD.”

See also

Examples

data(sim_miss)
impute_softimpute(sim_miss)
#>             X1           X2        X3           X4          X5         X6
#> 1  0.155050833  0.881156767 0.3509548 0.1194841210 0.600405471 0.66968808
#> 2  0.968378809  0.607484067 0.6337790 0.6181194666 0.008132938 0.75207300
#> 3  0.468263086  0.574529562 0.2476089 0.1272797810 0.370125689 0.54413862
#> 4  0.776819652  0.803133289 0.5513393 0.1459775318 0.724248714 0.84777606
#> 5  0.407885741  0.799116918 0.2347611 0.1897135850 0.418270750 0.70873497
#> 6  0.538797149  0.801862649 0.2586147 0.5754945197 0.238248518 0.62927422
#> 7  0.830082966  0.752073002 0.9528881 0.0001881735 0.887881226 0.48833525
#> 8  0.187103555  0.575461327 0.8568746 0.3283282877 0.577775384 0.34722847
#> 9  0.779969688  0.944218290 0.5195847 0.0782780354 0.739950257 0.20304801
#> 10 0.193943927  0.218987358 0.5545242 0.6041721753 0.160679622 0.73995026
#> 11 0.434231178  0.477997912 0.8769438 0.3267990556 0.529571799 0.89369752
#> 12 0.002274518  0.044668789 0.8018626 0.7904525734 0.574529562 0.57452956
#> 13 0.834692139  0.168485713 0.5467373 0.8300829662 0.960086967 0.08444870
#> 14 0.361093744  0.368280872 0.2327532 0.8100510929 0.398616886 0.14825833
#> 15 0.956967818  0.563311366 0.6022678 0.2027530000 0.408468068 0.37028128
#> 16 0.948497073  0.945908378 0.7125391 0.5060450307 0.613321482 0.86117762
#> 17 0.600729976  0.734327573 0.6489036 0.5206679751 0.706590850 0.83668180
#> 18 0.261807405  0.703656773 0.1875624 0.0842680801 0.799116918 0.74054705
#> 19 0.643034251  0.736824060 0.2383929 0.2687653562 0.161331222 0.98576580
#> 20 0.526233507 -0.008861609 0.8913933 0.8147157354 0.885932416 0.03208182