imputomics: The Biggest Collection of Methods of Missing Values Imputation for ‘Omics’ Data

Offers a range of methods for simulating and imputing missing values in different types of omics data such as genomics, transcriptomics, proteomics, and metabolomics. Provides tools for comparing and evaluating the performance of imputation methods in the terms of accuracy and computation time.

See also

Author

Maintainer: Michał Burdukiewicz michalburdukiewicz@gmail.com (ORCID)

Authors:

Krystyna Grzesiak krygrz11@gmail.com (ORCID) [contributor]

Dominik Nowakowski nowakowskisf@gmail.com (ORCID) [contributor]

Jarosław Chilimoniuk jaroslaw.chilimoniuk@gmail.com (ORCID) [contributor]

Jakub Kała kalajakub96@gmail.com [contributor]