MetabImpute GSIMP imputation.

A function to replace NA in the data frame by GSIMP.

Usage

impute_metabimpute_gsimp(missdf, verbose = FALSE, ...)

Arguments

missdf

a data frame with missing values to be imputed containing features in columns and samples in rows.

verbose

boolean, if TRUE, prints the typical prompts of MetabImpute::Impute().

...

other parameters of MetabImpute::Impute() besides method and data.

Value

A data.frame with imputed values of GSIMP by MetabImpute::Impute().

No replicates allowed

Impute() allows users to improve the quality of the imputation by providing the number of replications of the experiment. As one of the aims of the imputomics is to standardize the input and the output, our wrappers do not allow for this behavior.

References

Davis TJ, Firzli TR, Higgins Keppler EA, Richardson M, Bean HD (2022). “Addressing Missing Data in GC × GC Metabolomics: Identifying Missingness Type and Evaluating the Impact of Imputation Methods on Experimental Replication.” Analytical Chemistry, 94(31), 10912--10920. ISSN 0003-2700, doi:10.1021/acs.analchem.1c04093 .

See also

Examples

data(sim_miss)
impute_metabimpute_gsimp(sim_miss)
#>              X1           X2        X3           X4          X5           X6
#> 1  1.550508e-01 8.811568e-01 0.1233130 1.194841e-01 0.600405471 0.0005271873
#> 2  9.683788e-01 6.074841e-01 0.6337790 6.181195e-01 0.008132938 0.7520730018
#> 3  4.682631e-01 5.745296e-01 0.2476089 3.823232e-07 0.370125689 0.0032660007
#> 4  7.768197e-01 8.031333e-01 0.5513393 1.459775e-01 0.724248714 0.8477760591
#> 5  4.078857e-01 7.991169e-01 0.2347611 1.897136e-01 0.418270750 0.7087349731
#> 6  5.387971e-01 8.018626e-01 0.2586147 5.754945e-01 0.238248518 0.0316069257
#> 7  8.300830e-01 7.520730e-01 0.9528881 1.881735e-04 0.887881226 0.4883352455
#> 8  1.871036e-01 5.754613e-01 0.8568746 3.283283e-01 0.577775384 0.3472284742
#> 9  7.799697e-01 9.442183e-01 0.1830248 7.827804e-02 0.739950257 0.2030480122
#> 10 1.939439e-01 2.189874e-01 0.5545242 6.041722e-01 0.160679622 0.7399502567
#> 11 4.342312e-01 4.779979e-01 0.8769438 3.267991e-01 0.529571799 0.8936975205
#> 12 2.274518e-03 4.466879e-02 0.8018626 7.904526e-01 0.574529562 0.5745295617
#> 13 8.346921e-01 1.684857e-01 0.5467373 8.300830e-01 0.960086967 0.0844487038
#> 14 6.655214e-12 3.682809e-01 0.2327532 8.100511e-01 0.398616886 0.1482583326
#> 15 9.569678e-01 5.633114e-01 0.6022678 2.027530e-01 0.408468068 0.3702812819
#> 16 9.484971e-01 3.481341e-03 0.7125391 5.060450e-01 0.613321482 0.8611776216
#> 17 6.007300e-01 7.343276e-01 0.6489036 5.206680e-01 0.706590850 0.8366817986
#> 18 2.618074e-01 7.036568e-01 0.1875624 8.426808e-02 0.799116918 0.7405470510
#> 19 6.430343e-01 7.368241e-01 0.2383929 2.687654e-01 0.161331222 0.9857658015
#> 20 5.262335e-01 4.374617e-06 0.1804625 8.147157e-01 0.885932416 0.0320818219